video_pipeline/pyproject.toml
[project]
name = "video-pipeline"
version = "0.1.0"
description = "On-prem video pipeline POC (FastAPI + Kafka + Workers + MinIO + Postgres)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "boto3>=1.34.0",
  "aiokafka>=0.10.0",
  "psycopg2-binary>=2.9.9",
  "orjson>=3.10.0",
  "streamlit>=1.36.0",
  "requests>=2.32.0",
  "websocket-client>=1.8.0",
]

[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100

